WIP
Basic grid mostly done, fetch content not working still playing around with how I want to approach that. Sub routes with a appended /value can't reference resource includes with nested resources that call public assets in the more general context. So when rendering the nested footer within the content page is not rendering the icons in the footer. content/assets/...