Content Grid WIP #2
Content Gird
Fetch the article by Id, return article data, and populate the content section.
Adds Mutation Observer to watch for the content update then enact any needed Js on the newly added elements. This handles the issues with deferred event handling as we use our mutation events to enact our new element functionality. For example the close button for the article to re-show the content grid and currently clear the content section.