Add JSON/Html
You can hit any of the content endpoints with either a html path or json path to be served the two different formats of a given content page. We use the json version to remote server on a content grid. That fetch is then inline cached on the card so we don't need to make any additional requests for that content.